Move SimpleCard and add story
